fix: search page has been converted to the app page view
