+++ src/search.tsx
@@ -2,12 +2,8 @@ import type {SearchConfig} from './components/Search';
 
 import React from 'react';
 import {createRoot} from 'react-dom/client';
-import {Lang, configure as configureDocs} from '@diplodoc/components';
-import {configure as configureUikit} from '@gravity-ui/uikit';
 
-import {Search, SearchProvider} from './components/Search';
-import {RouterProvider} from './components/Router';
-import {LangProvider} from './hooks/useLang';
+import {SearchPage} from './components/Search';
 
 const root = document.getElementById('root');
 
@@ -15,30 +11,15 @@ if (!root) {
     throw new Error('Root element not found!');
 }
 
-const url = new URL(window.location.href);
-
-const windowData = window.__DATA__ as SearchConfig;
-const searchConfig = windowData || {};
-const searchData = windowData.search;
-
-const lang = (searchData?.lang as Lang) || Lang.En;
-
-configureUikit({lang});
-configureDocs({lang});
-
+const data = window.__DATA__ as SearchConfig;
+const searchConfig = data || {};
 searchConfig.depth = searchConfig.depth || 3;
 
-const routerConfig = {
-    pathname: url.pathname,
-    depth: 3,
+const url = new URL(window.location.href);
+
+const props = {
+    data,
+    url,
 };
 
-createRoot(root).render(
-    <LangProvider value={lang}>
-        <RouterProvider value={routerConfig}>
-            <SearchProvider value={searchData}>
-                <Search />
-            </SearchProvider>
-        </RouterProvider>
-    </LangProvider>,
-);
+createRoot(root).render(<SearchPage {...props} />);
